+++ src/App.css
@@ -12,6 +12,8 @@ body {
 }
 
 .main-content {
-    padding-top: 100px; /* 헤더 높이만큼 여백 추가 */
-    padding-bottom: 60px; /* Footer 높이만큼 여백 추가 */
+    display: flex; /* Flexbox 사용 */
+    min-height: calc(100vh - 160px); /* 전체 높이에서 패딩 값을 뺀 높이 */
+    align-items: center;
+    justify-content: center;
 }
