+++ src/pages/Login/component/LoginComponent.module.css
@@ -1,7 +1,7 @@
 input {
     width: 100%;
     margin: 10px 0;
-    padding: 12px;
+    padding: 12px 0px;
     border: 1px solid #ccc;
     border-radius: 5px;
     font-size: 16px;
