+++ src/pages/Signup/component/SignupComponent.module.css
@@ -1,5 +1,5 @@
 .container {
-    margin-top: 200px;
+    margin-top: 100px;
     display: flex;
     flex-direction: column;
     align-items: center;
@@ -10,4 +10,48 @@
     display: flex;
     flex-direction: column;
     align-items: center;
+}
+
+.formContainer {
+    width: 100%;
+    max-width: 400px;
+    padding: 20px;
+    border-radius: 8px;
+    box-shadow: 0 4px 10px rgba(0, 0, 0, 0.1);
+    background-color: #fff;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+}
+
+h1 {
+    margin-bottom: 20px;
+    font-size: 24px;
+    font-weight: bold;
+}
+
+.buttonContainer {
+    display: flex;
+    justify-content: center;
+    margin-top: 30px;
+    width: 100%;
+}
+
+
+.buttonContainer button{
+    width: 100%;
+}
+
+
+.signupContainer {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    margin-top: 10px;
+}
+
+.options {
+    display: flex;
+    justify-content: space-between;
+    width: 100%;
 }
